added SimpleForm.  added fields to ExampleForm.
